Slight refactor for results-related e2e tests
