+++ templates/experiments/basic/src/config.yaml
@@ -33,4 +33,4 @@ text: 'Enter your experiment description here.'
 tagline: 'Be a citizen scientist! Try this quiz.'
 duration: ''
 showResults: true
-resultsType: 'modelPrediction'
+resultsType: 'percentileRank'
