+++ templates/experiments/basic/src/e2e/expInfo.js
@@ -9,6 +9,8 @@ const expInfo = {
   archived: expConfig.archived,
   paused: expConfig.dataPaused,
   simulationMode: true, // Do all jsPsych plugins used have simulation mode?
+  showResults: expConfig.showResults,
+  resultsType: expConfig.resultsType,
   dataRows: 1,
 };
